+++ api/src/resources/dataset.rs
@@ -89,12 +89,19 @@ pub enum TimeResolution {
 pub enum Attribute {
     Labels,
     AttachmentPropertyTypes,
+    AttachmentPropertyNumAttachments,
 }
 
 #[derive(Debug, Clone, Serialize)]
 #[serde(tag = "kind", rename_all = "snake_case")]
 pub enum AttributeFilterEnum {
-    StringAnyOf { any_of: Vec<String> },
+    StringAnyOf {
+        any_of: Vec<String>,
+    },
+    NumberRange {
+        minimum: Option<usize>,
+        maximum: Option<usize>,
+    },
 }
 
 #[derive(Debug, Clone, Serialize)]
@@ -123,6 +130,7 @@ pub struct StatisticsRequestParams {
 pub enum OrderEnum {
     ByLabel { label: String },
     Recent,
+    Sample { seed: usize },
 }
 
 #[derive(Debug, Clone, Serialize)]