+++ cli/src/commands/get/comments.rs
@@ -4,6 +4,7 @@ use chrono::{DateTime, Utc};
 use colored::Colorize;
 use dialoguer::{Confirm, Input, Select};
 use log::info;
+use rand::Rng;
 use regex::Regex;
 use reinfer_client::{
     resources::{
@@ -122,6 +123,14 @@ pub struct GetManyCommentsArgs {
     #[structopt(long = "attachments")]
     /// Save attachment content for each comment
     include_attachment_content: Option<bool>,
+
+    #[structopt(long = "--only-with-attachments")]
+    /// Whether to only return comments with attachment metadata
+    only_with_attachments: Option<bool>,
+
+    #[structopt(long = "--shuffle")]
+    /// Whether to return comments in a random order
+    shuffle: Option<bool>,
 }
 
 #[derive(Debug, Deserialize)]
@@ -304,6 +313,8 @@ pub fn get_many(client: &Client, args: &GetManyCommentsArgs) -> Result<()> {
         recipients,
         senders,
         include_attachment_content,
+        only_with_attachments,
+        shuffle,
     } = args;
 
     let by_timerange = from_timestamp.is_some() || to_timestamp.is_some();
@@ -328,7 +339,8 @@ pub fn get_many(client: &Client, args: &GetManyCommentsArgs) -> Result<()> {
         bail!("Cannot use a label filter when `dataset` is not provided.")
     }
 
-    if !attachment_type_filters.is_empty() && dataset.is_none() {
+    if (!attachment_type_filters.is_empty() | only_with_attachments.is_some()) && dataset.is_none()
+    {
         bail!("Cannot use a attachment type filter when `dataset` is not provided.")
     }
 
@@ -362,6 +374,10 @@ pub fn get_many(client: &Client, args: &GetManyCommentsArgs) -> Result<()> {
         bail!("Cannot include attachment content when no file is provided")
     }
 
+    if shuffle.is_some() && dataset.is_none() {
+        bail!("Cannot shuffle data when dataset is not provided")
+    }
+
     let OutputLocations {
         jsonl_file,
         attachments_dir,
@@ -383,6 +399,17 @@ pub fn get_many(client: &Client, args: &GetManyCommentsArgs) -> Result<()> {
             attribute: Attribute::AttachmentPropertyTypes,
             filter: AttributeFilterEnum::StringAnyOf {
                 any_of: attachment_type_filters.to_vec(),
+            },
+        });
+    }
+
+    let mut only_with_attachments_filter: Option<AttributeFilter> = None;
+    if only_with_attachments.unwrap_or_default() {
+        only_with_attachments_filter = Some(AttributeFilter {
+            attribute: Attribute::AttachmentPropertyNumAttachments,
+            filter: AttributeFilterEnum::NumberRange {
+                minimum: Some(1),
+                maximum: None,
             },
         });
     }
@@ -436,6 +463,8 @@ pub fn get_many(client: &Client, args: &GetManyCommentsArgs) -> Result<()> {
         attachment_property_types_filter,
         messages_filter: Some(messages_filter),
         attachments_dir,
+        only_with_attachments_filter,
+        shuffle: shuffle.unwrap_or(false),
     };
 
     if let Some(file) = jsonl_file {
@@ -490,6 +519,8 @@ struct CommentDownloadOptions {
     user_properties_filter: Option<UserPropertiesFilter>,
     messages_filter: Option<MessagesFilter>,
     attachments_dir: Option<PathBuf>,
+    only_with_attachments_filter: Option<AttributeFilter>,
+    shuffle: bool,
 }
 
 impl CommentDownloadOptions {
@@ -502,6 +533,10 @@ impl CommentDownloadOptions {
 
         if let Some(attachment_types_attribute_filter) = &self.attachment_property_types_filter {
             filters.push(attachment_types_attribute_filter.clone())
+        }
+
+        if let Some(only_with_attachments_filter) = &self.only_with_attachments_filter {
+            filters.push(only_with_attachments_filter.clone())
         }
 
         filters
@@ -656,7 +691,13 @@ fn get_comments_from_uids(
             messages: options.messages_filter.clone(),
         },
         limit: DEFAULT_QUERY_PAGE_SIZE,
-        order: OrderEnum::Recent,
+        order: if options.shuffle {
+            OrderEnum::Sample {
+                seed: rand::thread_rng().gen_range(0..2_i64.pow(31) - 1) as usize,
+            }
+        } else {
+            OrderEnum::Recent
+        },
     };
 
     client